qa/cephfs: don't use int() to convert string of float point number
Fixes: http://tracker.ceph.com/issues/20582

+++ qa/tasks/cephfs/cephfs_test_case.py
@@ -159,7 +159,7 @@ class CephFSTestCase(CephTestCase):
 
         # Load an config settings of interest
         for setting in self.LOAD_SETTINGS:
-            setattr(self, setting, int(self.fs.mds_asok(
+            setattr(self, setting, float(self.fs.mds_asok(
                 ['config', 'get', setting], self.mds_cluster.mds_ids[0]
             )[setting]))
